utils/configdataset.m (config_imname, config_qimname): fix for dots in filename.
It is not possible to use `fileparts` to find if a string includes the
file extension of a filename.  `fileparts` is only useful to split a
filepath in its components but we need to know ahead of time if there
is a file extension.  This is because the basename itself may have
dots.

This commit removes the logic that handles imlist and qimlist with the
file extensions.  It now always expects groundtruth imlist to be the
basename, i.e., without file extension.